+++ jgltf-impl-v2/src/main/java/de/javagl/jgltf/impl/v2/CameraOrthographic.java
@@ -27,14 +27,14 @@ public class CameraOrthographic
      * (required) 
      * 
      */
-    private Float xmag;
+    private Double xmag;
     /**
      * The floating-point vertical magnification of the view. This value 
      * **MUST NOT** be equal to zero. This value **SHOULD NOT** be negative. 
      * (required) 
      * 
      */
-    private Float ymag;
+    private Double ymag;
     /**
      * The floating-point distance to the far clipping plane. This value 
      * **MUST NOT** be equal to zero. `zfar` **MUST** be greater than 
@@ -42,13 +42,13 @@ public class CameraOrthographic
      * Minimum: 0.0 (exclusive) 
      * 
      */
-    private Float zfar;
+    private Double zfar;
     /**
      * The floating-point distance to the near clipping plane. (required)<br> 
      * Minimum: 0.0 (inclusive) 
      * 
      */
-    private Float znear;
+    private Double znear;
 
     /**
      * The floating-point horizontal magnification of the view. This value 
@@ -59,7 +59,7 @@ public class CameraOrthographic
      * @throws NullPointerException If the given value is <code>null</code>
      * 
      */
-    public void setXmag(Float xmag) {
+    public void setXmag(Double xmag) {
         if (xmag == null) {
             throw new NullPointerException((("Invalid value for xmag: "+ xmag)+", may not be null"));
         }
@@ -74,7 +74,7 @@ public class CameraOrthographic
      * @return The xmag
      * 
      */
-    public Float getXmag() {
+    public Double getXmag() {
         return this.xmag;
     }
 
@@ -87,7 +87,7 @@ public class CameraOrthographic
      * @throws NullPointerException If the given value is <code>null</code>
      * 
      */
-    public void setYmag(Float ymag) {
+    public void setYmag(Double ymag) {
         if (ymag == null) {
             throw new NullPointerException((("Invalid value for ymag: "+ ymag)+", may not be null"));
         }
@@ -102,7 +102,7 @@ public class CameraOrthographic
      * @return The ymag
      * 
      */
-    public Float getYmag() {
+    public Double getYmag() {
         return this.ymag;
     }
 
@@ -118,7 +118,7 @@ public class CameraOrthographic
      * the given constraints
      * 
      */
-    public void setZfar(Float zfar) {
+    public void setZfar(Double zfar) {
         if (zfar == null) {
             throw new NullPointerException((("Invalid value for zfar: "+ zfar)+", may not be null"));
         }
@@ -137,7 +137,7 @@ public class CameraOrthographic
      * @return The zfar
      * 
      */
-    public Float getZfar() {
+    public Double getZfar() {
         return this.zfar;
     }
 
@@ -151,7 +151,7 @@ public class CameraOrthographic
      * the given constraints
      * 
      */
-    public void setZnear(Float znear) {
+    public void setZnear(Double znear) {
         if (znear == null) {
             throw new NullPointerException((("Invalid value for znear: "+ znear)+", may not be null"));
         }
@@ -168,7 +168,7 @@ public class CameraOrthographic
      * @return The znear
      * 
      */
-    public Float getZnear() {
+    public Double getZnear() {
         return this.znear;
     }
