+++ jgltf-impl-v2/src/main/java/de/javagl/jgltf/impl/v2/CameraPerspective.java
@@ -26,26 +26,26 @@ public class CameraPerspective
      * Minimum: 0.0 (exclusive) 
      * 
      */
-    private Float aspectRatio;
+    private Double aspectRatio;
     /**
      * The floating-point vertical field of view in radians. This value 
      * **SHOULD** be less than π. (required)<br> 
      * Minimum: 0.0 (exclusive) 
      * 
      */
-    private Float yfov;
+    private Double yfov;
     /**
      * The floating-point distance to the far clipping plane. (optional)<br> 
      * Minimum: 0.0 (exclusive) 
      * 
      */
-    private Float zfar;
+    private Double zfar;
     /**
      * The floating-point distance to the near clipping plane. (required)<br> 
      * Minimum: 0.0 (exclusive) 
      * 
      */
-    private Float znear;
+    private Double znear;
 
     /**
      * The floating-point aspect ratio of the field of view. (optional)<br> 
@@ -56,7 +56,7 @@ public class CameraPerspective
      * the given constraints
      * 
      */
-    public void setAspectRatio(Float aspectRatio) {
+    public void setAspectRatio(Double aspectRatio) {
         if (aspectRatio == null) {
             this.aspectRatio = aspectRatio;
             return ;
@@ -74,7 +74,7 @@ public class CameraPerspective
      * @return The aspectRatio
      * 
      */
-    public Float getAspectRatio() {
+    public Double getAspectRatio() {
         return this.aspectRatio;
     }
 
@@ -89,7 +89,7 @@ public class CameraPerspective
      * the given constraints
      * 
      */
-    public void setYfov(Float yfov) {
+    public void setYfov(Double yfov) {
         if (yfov == null) {
             throw new NullPointerException((("Invalid value for yfov: "+ yfov)+", may not be null"));
         }
@@ -107,7 +107,7 @@ public class CameraPerspective
      * @return The yfov
      * 
      */
-    public Float getYfov() {
+    public Double getYfov() {
         return this.yfov;
     }
 
@@ -120,7 +120,7 @@ public class CameraPerspective
      * the given constraints
      * 
      */
-    public void setZfar(Float zfar) {
+    public void setZfar(Double zfar) {
         if (zfar == null) {
             this.zfar = zfar;
             return ;
@@ -138,7 +138,7 @@ public class CameraPerspective
      * @return The zfar
      * 
      */
-    public Float getZfar() {
+    public Double getZfar() {
         return this.zfar;
     }
 
@@ -152,7 +152,7 @@ public class CameraPerspective
      * the given constraints
      * 
      */
-    public void setZnear(Float znear) {
+    public void setZnear(Double znear) {
         if (znear == null) {
             throw new NullPointerException((("Invalid value for znear: "+ znear)+", may not be null"));
         }
@@ -169,7 +169,7 @@ public class CameraPerspective
      * @return The znear
      * 
      */
-    public Float getZnear() {
+    public Double getZnear() {
         return this.znear;
     }
