+++ jgltf-model/src/main/java/de/javagl/jgltf/model/impl/Cameras.java
@@ -61,10 +61,10 @@ class Cameras
      * @param result The array storing the result
      * @return The result array
      */
-    static float[] computeProjectionMatrix(
-        CameraModel cameraModel, Float aspectRatio, float result[])
+    static double[] computeProjectionMatrix(
+        CameraModel cameraModel, Double aspectRatio, double result[])
     {
-        float localResult[] = Utils.validate(result, 16);
+        double localResult[] = Utils.validate(result, 16);
         
         CameraPerspectiveModel cameraPerspective = 
             cameraModel.getCameraPerspectiveModel();
@@ -74,9 +74,9 @@ class Cameras
         
         if (cameraPerspective != null)
         {
-            float fovRad = cameraPerspective.getYfov();
-            float fovDeg = (float)Math.toDegrees(fovRad);
-            float localAspectRatio = 1.0f;
+            double fovRad = cameraPerspective.getYfov();
+            double fovDeg = Math.toDegrees(fovRad);
+            double localAspectRatio = 1.0;
             if (aspectRatio != null)
             {
                 localAspectRatio = aspectRatio;
@@ -85,8 +85,8 @@ class Cameras
             {
                 localAspectRatio = cameraPerspective.getAspectRatio();
             }
-            float zNear = cameraPerspective.getZnear();
-            Float zFar = cameraPerspective.getZfar();
+            Double zNear = cameraPerspective.getZnear();
+            Double zFar = cameraPerspective.getZfar();
             if (zFar == null)
             {
                 MathUtils.infinitePerspective4x4(
@@ -100,14 +100,14 @@ class Cameras
         }
         else if (cameraOrthographic != null)
         {
-            float xMag = cameraOrthographic.getXmag();
-            float yMag = cameraOrthographic.getYmag();
-            float zNear = cameraOrthographic.getZnear();
-            float zFar = cameraOrthographic.getZfar();
+            double xMag = cameraOrthographic.getXmag();
+            double yMag = cameraOrthographic.getYmag();
+            double zNear = cameraOrthographic.getZnear();
+            double zFar = cameraOrthographic.getZfar();
             MathUtils.setIdentity4x4(localResult);
-            localResult[0] = 1.0f / xMag;
-            localResult[5] = 1.0f / yMag;
-            localResult[10] = 2.0f / (zNear - zFar);
+            localResult[0] = 1.0 / xMag;
+            localResult[5] = 1.0 / yMag;
+            localResult[10] = 2.0 / (zNear - zFar);
             localResult[14] = (zFar + zNear) / (zNear - zFar);
         }
         else
